requirements.txt
argparse==1.4.0
pandas==1.2.4
mysql-connector-python==8.0.25
